Add RestApi For Importing Excel File To The Database With Multi Value

+++ repository/src/main/java/com/ecommerce/repository/PriceRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface PriceRepository extends JpaRepository<PriceEntity,Long> {
-    Optional<PriceEntity> findByPrice(Long price);
+    Optional<PriceEntity> findByPrice(Double price);
 }
